bugfix chrome download target

+++ lsdir.cpp
@@ -351,7 +351,9 @@ class LSDirectory {
         }
 
        private:
-        void print_size() { printf("%llu", uint64_t(s.st_size)); }
+        void print_size() {
+            printf("%lu", s.st_size);
+        }
         void print_dirname() {
             for(int i = 0; i < dlen; ++i)
                 printf("%c", fp[i]);  // dirname
